+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+#####May 10th, 2015
+- Added level loading from file feature
 #####April 29th, 2015
 - Glitches Completely implemented:
 	- Hovering Glitch